# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Twaky API
  description: HTTP + SSE surface over the Twaky mission engine.
  version: 0.1.0
paths:
...
  /api/agents/{agent_id}:
    get:
      tags:
      - agents
      summary: Get Agent
      operationId: get_agent_api_agents__agent_id__get
      parameters:
      - name: agent_id
        in: path
        required: true
        schema:
          type: string
          title: Agent Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/agents/{agent_id}/default_prompt:
    get:
      tags:
      - agents
      summary: Get Default Prompt
      operationId: get_default_prompt_api_agents__agent_id__default_prompt_get
      parameters:
      - name: agent_id
        in: path
        required: true
        schema:
          type: string
          title: Agent Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultPromptResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    Agent:
      properties:
        id:
          type: string
          title: Id
        display_name:
          type: string
          title: Display Name
        role:
          type: string
          enum:
          - orchestrator
          - specialist
          title: Role
        system_prompt:
          type: string
          title: System Prompt
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
        temperature:
          anyOf:
          - type: number
          - type: 'null'
          title: Temperature
        effective_model:
          type: string
          title: Effective Model
        updated_at:
          type: string
          format: date-time
          title: Updated At
      additionalProperties: false
      type: object
      required:
      - id
      - display_name
      - role
      - system_prompt
      - model
      - temperature
      - effective_model
      - updated_at
      title: Agent
...
    DefaultPromptResponse:
      properties:
        system_prompt:
          type: string
          title: System Prompt
      additionalProperties: false
      type: object
      required:
      - system_prompt
      title: DefaultPromptResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError